lambda/parquet_to_rds/requirements.txt
boto3==1.34.0
pandas==2.1.4
pyarrow==14.0.2
psycopg2-binary==2.9.9







